menu refactored
